Connect to a database

Create a connection object to a PostgreSQL database using an ODBC connection. For details about setting up an ODBC connection, see this guide.

Usage

db_connect(connection_name)

Arguments

connection_name

The name of the connection to connect to. This is the name of the connection as it appears in Connections pane in RStudio.

Value

A listed_connection, which is a list containing an ODBC connection object to a PostgreSQL database, as well as a nested list of references to schemas, tables and views in the database.

Examples

if (FALSE) { # \dontrun{
db <- db_connect("shared")
} # }